requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
celery[redis]==5.4.0
redis==5.2.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pinecone-client==5.0.1
anthropic==0.40.0
openai==1.57.4
pygithub==2.5.0
gitpython==3.1.43
tree-sitter==0.24.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-java==0.23.5
tree-sitter-go==0.23.4
neo4j==5.26.0
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.28.1
sse-starlette==2.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
